+++ src/Plugin.php
@@ -11,7 +11,7 @@ class Plugin
      *
      * @var string
      */
-    const VERSION = '1.2.5';
+    const VERSION = '1.2.6';
 
     /**
      * The plugin slug.